python-backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
pydantic==2.10.3

# YOLO and ML
ultralytics==8.3.50
torch==2.9.1
torchvision==0.24.1

# Image processing
opencv-python==4.10.0.84
pillow==11.0.0
numpy==1.26.4

# CORS support
python-dotenv==1.0.1
